New target: WASI
not sure yet how useful that is, but it was easy to create: With
`-wasi-system-api` the output is suitable for running under `wasmtime`,
and it allows us to `print` (which is not the case with `-no-system-api`
and running under `wasm`). The main advantage is that we can write
better test cases in `test/run`, or more easily debug them.

So maybe we could make this the default?