Static wasm linker: update memory size
previously, we would leave the memory limits in place, but by adding the
`(data)` sections from the other module, might have data sections out of
range.

 * `wasm-validate` did not notice
 * V8’s validator did not notice (but we’d get this unhandled exception)
 * `wasm-opt` from the binaryen toolset gave me the right hint.

This bug was dependent on the number of static strings in the C RTS
code, hence the seemingly random occurrence of this problem in #687.