Try to get rid call_conv
the `call_conv` idea as part of `FuncE` was mostly misguided, at least
in the current form (with `FuncE` having a list of argument and return
types), as there is a redundancy.

Here I explore if we can get rid of it, or at least confine it to the
interpreter code (for now).

In principle yes, I think. In practice I have introduced bugs that I
will have to look at later.