Emulate ICReplyE/ICCallE on the old platform
This models the IC’s bookkeeping of the current reply-target:

* For promising functions, the backend adds another argument for the reply
  continuation.
* During execution, it keeps track of the reply continuation in a global
* When doing further calls, it uses `func.bind_ref` to bundle the
  continuation with the callbacks.

This requires the backend to distinguish promising functions from
others. It uses the control field (`Promises 0`). The list of return
types is still set to `[]`. This is probably a _bad hack_. In general,
the typing needs to be cleaned up for the new platform.

Maybe the `async` does not have to change any types, and we merely
adjust the typing rules in the presence of a flavor flag.

Callbacks are still shared functions, for now.

The IR interpreter cannot handle `ICCallPrim` yet. This simply disables
running the IR interpreter for now.

Not everything works yet.