The T.Func refactoring: Getting rid of as_seq
the goal of this PR is to get rid of `T.as_seq`, as almost any use of
that is a code smell: From the type `(t1,t2)` we cannot tell any more
whether this was meant to be a two-element sequence, or a one-element
sequence with a first-class tuple inside.

The core of the refactoring is to change the `T.Func` type to put the
async’s return _sequence_ into the `ret_tys` field. I.e.
```
T.Func (Shared, Promises n, tbs, ts1, [Async (Tup ts2)]) where n = length ts2
```
becomes
```
T.Func (Shared, Promises, tbs, ts1, ts2)
```

Note that there is much less redundancy in these types, much less
invalid combinations are possible.

A fair number of pattern matches on both `control` and the `return typ`
now becomes a simple `match` on the `control`, and other things become
easier. This is a good sign.

The `T.codom c ts` function can be used to easily construct the codomain
of the function (like `T.seq ts` for arguments); it adds `Tup` and
`Async` as needed.

The biggest wart is that the IR-check needs a different `T.codom` post
the `async` translation pass. Maybe we can clean that up later.

There is still a bit use of `T.as_seq` in `Ir_construct`; this might be
ok, it might just “opportunistically unbox tuples in arguments in
synthesized code”. And a use in `T.async`. Maybe one can get rid of it
if `T.Async` itself becomes n-ary.

But otherwise `T.as_seq` is gone. There is now a safer `T.seq_of_tup`
used in the desugarer when there is guaranteed to be a `T.Tup` in the
argument.