Add reversei test case
this is a test case based on
https://github.com/dfinity-lab/dapps/blob/75ead35363574f3697e37cd3a0592e51d3253a36/examples/reversi/src/reversi/main.mo
resp. https://github.com/dfinity-lab/dapps/pull/4

@ninegua reports in
https://dfinity.slack.com/archives/CPKAD5BJN/p1572573329177100
that there is an out-of-memory-condition, so I tried to reproduce it,
but I could not easily. Maybe it needs more calls or other code. But its
still a nice example.

@ninegua, if you remeber how to trigger the out-of-memory condition (if
it even was a Wasm/Motoko out of memory), maybe you can tell us or amend
the test case.