Build ocaml programs (moc,…) statically (#823)
On Linux, we can build ocaml binaries statically, so that the resulting binary works on all linux distributions, without any complex or fragile bundling into self-extracting binaries. So let’s do that
for release builds. Building in `nix-shell` is unaffected (this also means that running `make` in `src`
in a “normal” opam environment will work as before).

On Darwin, we use  `nix/standalone-darwin.nix` to make sure we fetch system libraries from, well,
the system.

This is copied from
https://github.com/dfinity-lab/common/blob/master/nix/overlays/standalone/darwin.nix
It is copied verbatin, so easy to update, but does not couple the
repositories (important for eventual open sourcing)