Class that tracks the history of a variable
This is code in anticipation of canisters that need to track the history of changes made to a variable and maintain that history in their current state,
e.g., for audit purposes.

Current sample code uses arrays, not lists, which may be wasteful but should be enough to play with applications.

The changes are tracked along with the times when changes occurred. For this, a mock module called "environment" is used in place of a real sys calls to obtain the time.