New flag -hide-warnings
The main purpose of this is to make our test suite less nosiy. Every
test case is first type-checked and then run in various modes. If there
are type checker warnings, but no errors (usually due to partial
patterns, I am looking at you, @crusso ;-)), we get these warnings in
the test suite output many times. This flags allows us to make sure that
running/compiling files is actually silent.

This is in particularly useful as empty output file are not even
stored by the test runner, so we have less files. Which is good.