Pass function and actor references as data
this implements the design in https://github.com/dfinity-lab/motoko/pull/883

For the forseeable future, actor references are `Blob`s, and function
references are `(blob, text)` tuples. So this uses these types in the
IDL. This allows us to pass any kind of reference we have.

Important step to get rid of the `dvm` dependency.