Pass function and actor references as data, try two
the previous approach was too hacky, so this is better: We introduce a
type `Blob` for packed arrays of word8. This is, so far, only used in
the backend, but it might come handy anyways (e.g. should we want to
pull the IDL encoding/decoding into an IR pass).

This does not update the Motoko-to-IDL export, which is actually tricky;
see discussion in #913.