Collect RTS stats (#925)
* Collect RTS stats

namely:
 * current size of the _dynamic_ heap
 * total number of bytes ever allocated
 * number of outstanding callbacks

One testcase checks that allocating a 10000 byte array allocates 10008
bytes.

The other checks that callbacks are freed as they are returned. Which
currently is not happening yet, but I will work on that next (test
driven development).

To be able to test that I am allowing first-class asyncs on the stub
API. It is not IR-type-safe yet until
https://github.com/dfinity-lab/motoko/pull/860 lands.

* Update test output

* Update src/codegen/compile.ml

Co-Authored-By: Claudio Russo <claudio@dfinity.org>