Pass function and actor references as data (#913)
this implements the design in https://github.com/dfinity-lab/motoko/pull/883

For the forseeable future, actor references are `Blob`s, and function
references are `(blob, text)` tuples. So this uses these types in the
IDL. This allows us to pass any kind of reference we have.

An important step to get rid of the `dvm` dependency.

This also  introduces a type `Blob` for packed arrays of word8. This is,
so far, only used in  the backend, but it might come handy anyways
(e.g. should we want to pull the IDL encoding/decoding into an IR pass).

This does _not_ update the Motoko-to-IDL export, which is actually tricky;
see discussion in #913.