Compile async expression to SelfCallE (#931)
so that we can have `async {…}` on the new platform! Yay!

Collaborative work with Claudio.

Introduce SelfCallE (not a prim, because not call-by-value) is essentially
a short-hand for a combination of `FuncE Shared` with immediate `CallE`.
It has to be kept togehter for the backend, which can compile this combination,
but not a general `FuncE Shared` expression.