Drop support for the M1 client
and no longer depend on `dvm` or the `dev` repo.

so much work goes down the drain… fake orthogonal persistence… trying
hard to avoid boxing and unboxing references… the `ElemTable`… support
for the DFINITY-specific custom sections. 800 lines of dense code.

But yay, finally nothing is using `dvm` any more, or importing stuff
from https://github.com/dfinity-lab/dev! So we only need that repo to
read about how the Nix cache works in the README…

The main feature that we cannot test any more now are non-top-level
shared functions. But most aspects of that code are still present in
some form, and can be re-arranged later.