Garbage collect closures (#944)
This defines a suitable data stucture in the C code (a packed array with
integrated free list). It is compatible with the RTS's idea of an array,
so it can be copied and traversed by the GC.

There is a test case that checks that that closed-over values (a large
array in this case) are indeed freed. We can add such test cases for
other GC-related things now.

There is also a unit test of the C data structure in test_rts.c.