Test suite hack: test actors everywhere
to consolidate the following

* we want to run tests including actors on drun/stub
* drun only supports a single top-level actor
* we want to run something during tests, including remote calls and
  await.
* the language does not allow await from the the program or actor
  top-level
* the system does not even allow function calls from actor
  initialization (which also rules out `ignore async { … }`)
* the interpreter runs programs which may create an actor and then use
  it.
* the compiled targets can invoke public methods

this proposes the following hack. Tests take the following form:

```
actor a {
  … some exports …
  public func go() = ignore async {
    … the test driving code …
  };
};
a.go(); //OR-CALL ingress go "DIDL\x00\x00"
```

The interpreter can run such a thing directly.

Before compilation, `run.sh` rewrite the last line to
```
//CALL ingress go "DIDL\x00\x00"
```
Now the test is a single top-level actor, thus it can be compiled to a
canister, and the `//CALL` comment is picked up by `drun-wrapper.sh` to
invoke `go()`.

All pretty ugly, but do we have better alternatives?