Split various test directories into subderivations (#954)
* gitSource.nix: Use cleanSourceWith

following the example in
https://github.com/dfinity-lab/dfinity/commit/4f2d90887218b1a243f3f21c74b0d2d967b748ad

This changes changes the derivations used by `cabal2nix` to only contain
the `.cabal` files. Which is a good thing.

* Merge test/run-release into test/run

with a new
```
+//MOC-FLAG --release
```
feature (may be handy for other things too).

* Make test/repl self-contained

by copying `type-lub.mo`, but probably fine.

* Split various test directories into subderivations

having spent so much time on nix this week I felt motivated to refine
our `test` derivation, by splitting it up into several subtests.

The benefits are
 * more parallelism
 * less rereuns (e.g. bumping `drun` will only re-run the tests using
   it)
 * quicker feedback on GitHub, e.g. the fact _which_ subderivation has
   failed will already be useful feedback.

* Fix shell and all-systems-go