Expose Blob type, with equality and comparisons
this does not yet add intro forms. The goal of this is that we can
merge and play around with #789 (caller information in the context)
while we flesh out questions like

 * what type is the caller? `Entity`? Do we want users and canisters
   separate? etc.

 * do we want to have `type Blob = [Word8]`? What does that mean for
   comparisons? Can we make that efficient?

Without intro forms we can’t test much yet. Maybe we can add intro
forms, but that’s suitable for a separate PR.