Remove dependency on the SDK repo
according to Hans on Slack, they are testing the JS user library there,
so we don’t have to, and these cross-repo dependencies are bothering
some people who have to deal with the CI implications. So lets get rid
of this.

We still check that the output of `didc` is valid JS using node, and we
any unintentional changes to the output will be flagged by the test
suite.