DFX Interface: Canister data
this is a concrete proposal for how to let dfx know Motoko about the
ids and types of canisters. The highlevel idea is:

 * dfx gets a list of all imported canisters (via #979). canisters can
   be referred to by their concrete id or an alias.
 * dfx provides the concrete id and the type (however it got that) in a
   file where motoko finds it

That file uses the `actor "ic:ABCDE01A7"` form proposed by @rossberg in #667.

This proposed to use `import A "ic:…"` to import actors, in contrast to
`import A "mo:pkgs/file"` which imports modules from packages. If using
the URL scheme to distinguish the two is not approporate, an alternative
would be
```
actor import A "ic:…"
```
which nicely rhymes with other actor-prefixed things like “actor class”.

Another open question is if it is fine if these actor files form their
own form of importable file (bare actor expression), or if they must be
modules. But if they are modules, then what name has the (single) member
that contains the reference? And would the importing user have to
actually project it?

Maybe `import A "ic:name"` could be syntactic sugar for
```
import {actor = A} "mo:_canisters/name"
```
and reduce the whole feature to the existing
importing-modules-from-canisters feature