Ropy texts: Introduce concat heap operator
and update operations in `text.c` to use it.

A test case checks that we can now build huge strings based on sharing.
The implementation enforces a hard limit of 2^30 bytes, though.

The code uses stack space liberally. I am not sure how much of a problem
that is. If it is, we might have to revisit some of the algorithms.