Tighten actor check on ICMode (#996)
* Do the show-translation per actor

there was a long-standing bug that using `debug_show` in a close actor
would make it non-closed, and @crusso just hit it. So I guess I should
have fixed that.

One solution is to run the show translation independently in the main
program and each actor.

@crusso, it seems that the detection of `actor` literatls in ICMode in
`typing` is incomplete.

* Ugh, stupid bug. Should have just shadowed :-)

* Tighten actor check on ICMode

should be merged after #994.

* Newlines

* Update test