Test suite: Run in both `drun` and `ic-stub`
I am not quite sure what the right thing to do is: Shall we run all
tests, unless otherwise configured, in both `drun` and `ic-stub`, so
that when the output changes it is clearer whether the problem is likely
with us or `drun` or the stub?

Or is it just wasteful and noisy?

Concretely, enabling `async { … }` on `drun` makes some of our tests
less useful, as `drun` does not wait for all queus to drain; it only
runs until a response appears, which can be early. By running tests in
both we can enable `async` on the ic, while still getting the more
thorough testing from `ic-stub`.

What do people think? @crusso @ggreif?