IR: Introduce lexp (#1009)
* IR: Introduce lexp

I was never quite happy about the partial functions `compile_lexp` and
`c_assign`, and when working on #1004, this got in the way too.

So let's be more honest and create a data type `lexp` for the kind of
things that may appear on the left of `:=`, as I think we dicussed on
some slack discussion once.

* Traverse all annotations in async.ml

* Apply suggestions from code review

Co-Authored-By: Claudio Russo <claudio@dfinity.org>

* Simplify Check_ir.check_lexp