Add support for caller ID
* add support for context arguments  as optional patterns on shared functions

<sort_pat>  = shared query? <pat>?

* implement context arguments in interpreters and backend

In a separate PR, I'd like to go further and add context arguments not just to the function space but all term interpreters so we can separated things which are closed over in environments and the ambient context, which is not. That should let us more accurately model the system API by keeping not just the caller but the reply and reject continuations in the context, with IC prims examining the context, not environment, to retrieve them.

We might even be able to detect reply to wrong caller by comparing the reply in the context against a ghost reply stored in the environment, using either physical equality (cheap and cheerful) or a stamped value.