Test suite: Run in both `drun` and `ic-stub` (#1018)
This runs all tests,  unless otherwise configured, in both `drun` and `ic-stub`, so
that when the output changes it is clearer whether the problem is likely
with us or `drun` or the stub. This would have caught #1036.

Also , enabling `async { … }` on `drun` makes some of our tests
less useful, as `drun` does not wait for all queus to drain; it only
runs until a response appears, which can be early. By running tests in
both we can enable `async` on the ic, while still getting the more
thorough testing from `ic-stub`. So this unblocks #958.

And once we open source, we can make the testsuite ignore
the `drun` tests if no `drun` is available. People won’t be able to run
`make accept`, but at least they can run some of the tests.