Ropy texts: Introduce concat heap operator (#985)
* Create a RTS text.c module (imported, but unused)

* stash

* Use basically all text operations into text.c

which removes a lot of hairy Wasm code from the backend, and also opens
the path towards smarter implementations in `text.c`

* The null byte is valid utf8

hat tip to @ggreif’s randomized tests that caught this.

* Constant, not linear

* Ropy texts: Introduce concat heap operator

and update operations in `text.c` to use it.

A test case checks that we can now build huge strings based on sharing.
The implementation enforces a hard limit of 2^30 bytes, though.

The code uses stack space liberally. I am not sure how much of a problem
that is. If it is, we might have to revisit some of the algorithms.

* Update rts/text.c

Co-Authored-By: Claudio Russo <claudio@dfinity.org>

* Update rts/text.c

Co-Authored-By: Claudio Russo <claudio@dfinity.org>

* Update rts/text.c

Co-Authored-By: Claudio Russo <claudio@dfinity.org>

* Typos

* Update rts/text.c

Co-Authored-By: Gabor Greif <ggreif@gmail.com>

* Update rts/text.c

Co-Authored-By: Gabor Greif <ggreif@gmail.com>

* Update rts/text.c

Co-Authored-By: Gabor Greif <ggreif@gmail.com>

* Update rts/text.c

Co-Authored-By: Gabor Greif <ggreif@gmail.com>

* Move alloc_blob into rts.c

* Update rts/text.c

Co-Authored-By: Gabor Greif <gabor@dfinity.org>

* Apply suggestions from code review

Co-Authored-By: Gabor Greif <gabor@dfinity.org>