compile_local_function: Do not use `VarLoc.deferred`
The `Varloc.deferred` stuff in the backend is quite an ugly hack, and I
want to get rid of it.

This gets rid of one of the two uses of it.

(I know that the method names in the `VarEnv` are kinda unhelpful,
because there `local` may refer to “local Mokoto variable” and “Wasm
local”.)