IR: Remove self binder from ActorE
instead,
 * add a `SelfRef` prim
 * in the desugarer, add a binding to the body of the Actor if needed

The “if needed” is a bit ugly, because of ubiqious recursion and that
`ActorE` is like a block with a single recursive group. So we only add
it if it is not shadowed.

This still lacks support for `SelfRef` in the IR interpreter, which is
blocked on support for `ActorOfIdBlob`, I believe. @crusso, what is the
status there? Is that easy to implement? I guess it needs the “Actor
heap”?