Check_ir: No `gather_scope`
in another PR, I had to remove this to make it typecheck, and removing
it does not break any types.

It also doesn’t make sense to me: I would expect
```
{
  type T = Int;
  let x = {} : {type T = Int;}
  foo () : T = 5;
}
```
be legal (not a duplicate definition as before) and
```
{
  let x = {} : {type T = Int;}
  foo () : T = 5;
}
```
illegal, because `T` is not in scope; `x.T` is, (I believe that the
current code accepts this).

@crusso, please double check.