AllocHow.decs: Also do fixed-point iteration for top-level blocks (#1099)
* AllocHow.decs: Also do fixed-point iteration for top-level blocks

Previously, the code was taking some short-cuts for top-level blocks,
and that was fine. But I want more than just functions to be static, and
the two different implementations here were a problem. So this unifies
them again, making just surgical adjustments where needed.

No change to any wasm output.

Adds a test case that I broke along the way.

* Update src/codegen/compile.ml

Co-Authored-By: Gabor Greif <gabor@dfinity.org>

* Update src/codegen/compile.ml

Co-Authored-By: Gabor Greif <gabor@dfinity.org>

* fix compilation

Co-authored-by: Gabor Greif <ggreif@gmail.com>