IR Interpreter: Support actor ids
this changes the value representation of an `actor …` to use `Text`,
i.e. a blob. This matches the RTS.

When accessing an actor field, we look up the actor in a global actor
map (representing, in a way, the state of the Internet Computer).

We trap more eagerly upon missing actors or fields than the platform,
but this is ok.

This implemented `ActorOfIdBlob` (now trivial), and unblocks #1104.