Create static heap objects for constant values on demand
Previously, with code like
```
func foo() = …;

foo();
higher_order(foo);
higher_order(foo);
```

we would initially not create a heap representation for `foo`, and just
do a direct call in `foo()`.

But what if we _do_ need a heap object for `foo`? We would allocate it
_dynamically_ upon each use, i.e. twice in the code above!

With this change, if there is any first-class use of `foo`, we put the
corresponding heap object into the static memory once, and use it in
every use of `foo`.

If there is no first-class use of `foo`, then no memory will be used.

This will be more important with #1121, because we really don't want to
synthesize whole module objects dynamically.

The testcase confirms the improvement.