GC: Only traverse mutable static heap objects
previously, the GC would scan the whole static heap for roots. This is
wasteful, as it includes immutable static objects (e.g. strings) that
cannot contain static roots.

With this change, the RTS emits, to static memory, an array of roots
with all mutable elements in the static memory, and traverses only
these.

The provided test case breaks if we forget to add the mutable variable
there to the list (but of course there is a risk of hard-to-debug bugs
if in the future we add something there).

Can be sped up further by a version of `Array.idx` that does not do
bounds checking, which can be used in other places as well.