nix/sources.json: use the same nixpkgs as dfinity & sdk
This is an experiment to see if motoko builds with this revision of
nixpkgs. It's needed because this nixpkgs builds on our new zh-hydra
which has access to the upstream cache.nixos.org disabled.

This setting has the advantage that our cache gets populated with
everything needed to build DFINITY. Secondly we reduce an attack
vector where an attacker can install a trojan in the cache.

If motoko really requires nixpkgs master than we can cherry-pick the
needed patches on top of nixpkgs master.