nix/sources.json: upgrade nixpkgs to the latest master (#1189)
This upgrades `nixpkgs` to DFINITY's fork of nixpkgs which tracks the
`master` branch. There're only two patches on top of `master`:

```
commit a4e23f32f8d7006e217125a9123b53597b9c0c24 (HEAD -> dfinity-motoko, dfinity-lab/dfinity-motoko)
Author: Bas van Dijk <v.dijk.bas@gmail.com>
Date:   Wed Feb 12 16:08:05 2020 +0100

    expect: remove autoreconfHook since that causes a wrong path to stty

    See: https://github.com/NixOS/nixpkgs/issues/79863

commit f359aea057db81c51e55d7250283a0b2201bd6b6
Author: Bas van Dijk <v.dijk.bas@gmail.com>
Date:	Thu Feb 6 19:14:09 2020 +0100

    libuv: 1.34.1 -> 1.34.2

    (cherry picked from commit 21ad5c123b8e76ab241b89acbc8db24925244d63)
```

We need this patch in order to build `motoko` without using the
upstream `cache.nixos.org`. It would be good not to use an external
cache to decrease the chance an attacker can infect the cache with a
trojan and to ensure DFINITY can always build their own code without
depending on an external cache.