Test runner: Check if `drun` and `ic-stub-run` are installed
if they are not, warn, but do not try to run the tests, and do not allow
accepting tests.

This means that people who do not have access to `dfinity` can run some
tests (including a public CI like Github Actions).

It also means that `make accept` will not commit `could not find
ic-stub-run` as the expected output, as has happened to us a few times
before.