Nix: Include sources.nix in repo
the fact that since #1113, the `sources.nix`, which is central to any
nix evaluation, is fetched remotely almost prevented me from doing work
offline. (Luckily I knew about `--option tarball-ttl 100000`).

But even offline, it seems silly for nix to check github every hour if
this file has changed (it will not, because it is pinned using a git
revision.) Copying to the repo seems to make more sense.

(Alternatively, it would make sense to fetch it in a way that specifies
the expected file hash, so that nix will download it exactly once, maybe
using `builtins.path` somehow.)

Unfortunately, other repos that we pull in still fetch that file
remotely, so you still need to pass `--option tarball-ttl 100000` to
work offline. I’ll use this PR to notify our nix gurus of this issue.