bring back builtin_types (these are listed for reference)
normally types get emitted into the bulk tag (0) on demand
but at the end we werge those into the unit tag.

We have to ensure that tags in the unit (these are types that are not
mentioned by the bulk tags yet, included in DWARF for completeness)
that contain references are written _after_ the tags they reference.

Note that the list is built in a reversed fashion.