Makes dfx pass arguments via a file
Instead of passing them as CLI arguments, this would allow the language
service to watch that file and respond to changes. The current
interface (using CLI arguments) is incompatible with a long-running
process like the language service, because it needs to react to a
changing environment. Canisters or packages can be added, renamed,
moved, or deleted while the language service is running.