IR: No shared notes
The `note` field of `Ir.exp` is mutable. This will be very handy as we
add pure analysis passes, i.e. passes that traverse the AST (but don’t
change it) and just want to set these notes.

But this is only viable if these notes are never shared. So this adds an
IR check that will complain if such notes are shared (or if `Check_ir`
visits a node twice, which would also be bad).

Hunting down all the cases where we currently shares nodes is tricky.
Maybe we need to restructure `construct.ml` to have a different type for
variables after all, so that we can `clone` in `varE`.