Check_ir: Check the const annotation
this is in preparation for the const-analysis-pass (which I’ll do in a
separate PR): It adds the `const` to the notes, and checks that they are
correct in Check_ir.

I wonder if `check_exp` in `Check_ir` should be refactored into the
traversal (does no checking, only updates `env`) and the checks
(non-recursive).