Compile: More disciplined detection of the TopLevel
the top level is where every `FuncE` is constant, because every possibly
captured value is either contant, or if not, it can be put in a
well-known static location.

Not top-level are: bodies of functions, and loops.

This actually simplifies some of the code, while simulatenously
producing better output (more constant values).