Const: Ignore top-level-bound free variables for FuncE
the analysis was not precise enough: a `FuncE` is constant if every free
variable is either
 * constant (duh)
 * or top-level bound

Rationale: To compile a function statically (without a closure) we just
need to know the _location_ of its free variables, but not their values.
And for all top-level things we will know where they end up.

With this change, actually _using_ the information (#1336) will produce
identical code as before, so hopefully this means that have not
regressed in any way.