Backend: Static bigint constants (#1357)
Previously, when we had a `Nat` literal that cannot be represented as an
unboxed scalar, we constructed it by creating multiple
machine-word-sized literals, and adding them together using the
libbtommath functions. The reason was that this would respect the
abstraction layer introduced by that library.

But the downside is of course kinda stupid run-time calculations. Worse,
it stands in the way of making all literals constant.

So this fixes this; now the backend needs to know the structure of
`mp_int` (but it already kinda has to for GC purposes). I left comments
to the relevant definitions from `libbtommath.h`.

This is a spin off of #1336.