Compile: proper compile_lit_pat (#1366)
In the previous coode, `compile_lit_pat` was calling `snd (compile_lit
…)`, clearly throwing away information. This works fine if `compile_lit`
has precisely the shape that it had when `compile_lit_pat` was written,
but easily breaks when `compile_lit` changes.

And surely enough, this just bit me with hard-to-debug invalid Wasm.

So let’s respect the abstraction of `compile_lit` and declare which
stack representation we want here, using `compile_lit_pat`.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>