Backend: PublicMethod is not Const
right now, a value that references a public method on ourselv is _not_ a
`Const`, because we need to run code (`get_self_reference`) to figureout
the value.

This blocks adding const arrays, as that relies on the intended
behaviour that any kind of `Const` in static memory.

Also makes sure that `const.ml` takes this into account.

One could make these `Const` by introducing a special sentinel value for
the self-actor … but not now.