Backend: Introduce Const.lit (#1364)
* Backend: Static bigint constants

Previously, when we had a `Nat` literal that cannot be represented as an
unboxed scalar, we constructed it by creating multiple
machine-word-sized literals, and adding them together using the
libbtommath functions. The reason was that this would respect the
abstraction layer introduced by that library.

But the downside is of course kinda stupid run-time calculations. Worse,
it stands in the way of making all literals constant.

So this fixes this; now the backend needs to know the structure of
`mp_int` (but it already kinda has to for GC purposes). I left comments
to the relevant definitions from `libbtommath.h`.

This is a spin off of #1336.

* Backend: Introduce Const.lit

This is a refactoring towards proper constant literals, but should
not have an effect on its own.

Essentially it introduces Const.lit to factor through the compile paths
for “literal → stack” and “literal → static memory”. See comment at the
beginning of the `Const module`.

* Some cleanup

* Unify function names

* Apply suggestions from code review

* Update src/codegen/compile.ml

Co-Authored-By: Gabor Greif <gabor@dfinity.org>

* Make sentence structure more clear

* Update src/codegen/compile.ml

Co-Authored-By: Gabor Greif <gabor@dfinity.org>

* Update src/codegen/compile.ml

Co-Authored-By: Gabor Greif <gabor@dfinity.org>

Co-authored-by: Gabor Greif <gabor@dfinity.org>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>