CI: Allow merging PRs even before darwin finishes (#1384)
Running `drun` on darwin ins kinda slow, so often this is what's holding
up swift merging of approved PRs.

It’s been a while since we actually had a failure on drawin that we
didn’t also have on Linux.

With this change, the darwin jobs are still run (and we would likely see
failures), but we exclude them from `all-systems-go` for PRs. This way,
mergify will be able to merge an approved PR earlier, which makes
breaking up PRs into chunks more convenient.