CI: Detect and report if a PR does not affect Wasm output (#1387)
* CI: Detect and report if a PR does not affect Wasm output

something that I wanted for a long time, and also @ggreif was asking for
it: A PR that does not change the wasm output on any (compiling) file in
`test/*.mo` will repor that this is the case. Only if it is not the
case it will report the perf (size and gas) changes.

* Update ci-pr.nix

Co-Authored-By: Gabor Greif <gabor@dfinity.org>

* Update ci-pr.nix

Co-Authored-By: Gabor Greif <gabor@dfinity.org>

* Sneak in benign change

* fix test

Co-authored-by: Gabor Greif <gabor@dfinity.org>