IDL/RTS: Make find_field nicer
see comment about what it does now.

One nice effect is that we get nicer error messages if a field is
missing!

A prior version of this had title “Make find_field much nicer”, but that
failed to materialize as the type arguments are not fixed width, so
can’t use a fixed “end” pointer to indicate when we are done, but still
need the mutable counter.

This should make #1388 much easier to implement.