Const: Resolve object patterns statically (#1386)
this addresses an issues that @crusso pointed out in some earlier
review: We should allow let-declarations that bind more than just a
simple variable to be static. This is imortant for idioms like

    let {foo; bar} = some_module;
    foo()

to be compiled properly.

So far, only _irrefutable_ paterns are taken into account; refutable
static patterns would require implementing static equality checks for
all `lit` values and their corresponding `const` values, and we would
have to produce code that just traps anyways… so not for now.