CI: Allow merging without QC
this is similar in spirit as #1384: This configures the `tests.qc`
derivation to still run always, but not block merging PRs. It is usually
the longest-running job, and while it is a useful check, I don’t think
merging PRs will have to wait for the merge commit to pass this check.

We can revert this if it turns out to cause problems, of course.