test/compare-wat.sh: Also mangle files
this hack that files in `test/run-drun` work with the interpreter, but
properly compile only after mangling of `//OR-CALL` rears its ugly head
again – it made the `compare-wat.sh` script useless on these files (it
would just compare the error message you’d get otherwise).

So we have to mangle here as well.

Is there a way to not have to mangle? Maybe some hack in the interpreter
that runs public functions called `go()` or something?