Report Array out of bounds even on absurd index (#1440)
* Report Array out of bounds even on absurd index

Previusly, array index with absurd sizes would trap with “bigint error”
due to not being able to convert the bigint number to 32bits. This is
technically correct, but not helpful.

So we now handle that case and give the same error message. This fixes #1421.

* duh

* Pass desired error message to `bigint_to_word32_trap_with`

* Update test cases

* Update more test output

* Apply suggestions from code review

Co-authored-by: Gabor Greif <gabor@dfinity.org>

* Pass error message to `to_word32_with` in a single word

to shave off some extra instructions

* Forgot this function

Co-authored-by: Gabor Greif <gabor@dfinity.org>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>